Small source code cleanup - bad mixture of tabs & spaces made file hard to read.

+++ sendfax/sendfax.c++
@@ -62,8 +62,7 @@ sendFaxApp::sendFaxApp()
 
 sendFaxApp::~sendFaxApp()
 {
-    if (stdinTemp != "")
-	Sys::unlink(stdinTemp);
+    if (stdinTemp != "") Sys::unlink(stdinTemp);
     delete db;
 }
 
@@ -88,141 +87,142 @@ sendFaxApp::run(int argc, char** argv)
     SendFaxJob& proto = getProtoJob();
     db = new FaxDB(tildeExpand(dbName));
     while ((c = Sys::getopt(argc, argv, "a:b:B:c:C:d:f:F:h:i:I:k:M:P:r:s:t:T:V:x:y:12lmnpvwDENR")) != -1)
-	switch (c) {
-	case '1':			// restrict to 1D-encoded data
-	    proto.setDesiredDF(0);
-	    break;
-	case '2':			// restrict to 2D-encoded data
-	    proto.setDesiredDF(1);
-	    break;
-	case 'a':			// time at which to transmit job
-	    proto.setSendTime(optarg);
-	    break;
-	case 'b':			// minimum transfer speed
-	    proto.setMinSpeed(optarg);
-	    break;
-	case 'B':			// desired transmit speed
-	    proto.setDesiredSpeed(optarg);
-	    break;
-	case 'C':			// cover page: template file
-	    proto.setCoverTemplate(optarg);
-	    break;
-	case 'c':			// cover page: comment field
-	    proto.setCoverComments(optarg);
-	    break;
-	case 'D':			// notify when done
-	    proto.setNotification("when done");
-	    break;
-	case 'd':			// destination name and number
-	    addDestination(optarg);
-	    break;
-	case 'E':			// disable use of ECM
-	    proto.setDesiredEC(FALSE);
-	    break;
-	case 'F':			// override tag line format string
-	    proto.setTagLineFormat(optarg);
-	    break;
-	case 'f':			// sender's identity
-	    setFromIdentity(optarg);
-	    break;
-	case 'h':			// server's host
-	    setHost(optarg);
-	    break;
-	case 'I':			// fixed retry time
-	    proto.setRetryTime(optarg);
-	    break;
-	case 'i':			// user-specified job identifier
-	    proto.setJobTag(optarg);
-	    break;
-	case 'k':			// time to kill job
-	    proto.setKillTime(optarg);
-	    break;
-	case 'l':			// low resolution
-	    proto.setVResolution(98.);
-	    break;
-	case 'M':			// desired min-scanline time
-	    proto.setDesiredMST(optarg);
-	    break;
-	case 'm':			// medium resolution
-	    proto.setVResolution(196.);
-	    break;
-	case 'n':			// no cover sheet
-	    proto.setAutoCoverPage(FALSE);
-	    break;
-	case 'N':			// no notification
-	    proto.setNotification("none");
-	    break;
-	case 'p':			// submit polling request
-	    addPollRequest();
-	    break;
-	case 'P':			// set scheduling priority
-	    proto.setPriority(optarg);
-	    break;
-	case 'r':			// cover sheet: regarding field
-	    proto.setCoverRegarding(optarg);
-	    break;
-	case 'R':			// notify when requeued or done
-	    proto.setNotification("when requeued");
-	    break;
-	case 's':			// set page size
-	    proto.setPageSize(optarg);
-	    break;
-	case 't':			// times to retry sending
-	    proto.setMaxRetries(atoi(optarg));
-	    break;
-	case 'T':			// times to dial telephone
-	    proto.setMaxDials(atoi(optarg));
-	    break;
-	case 'v':			// verbose mode
-	    verbose++;
-	    setvbuf(stdout, NULL, _IOLBF, BUFSIZ);
-	    SendFaxClient::setVerbose(TRUE);	// type rules & basic operation
-	    FaxClient::setVerbose(verbose > 1);	// protocol tracing
-	    break;
-	case 'V':			// cover sheet: voice number field
-	    proto.setCoverVoiceNumber(optarg);
-	    break;
-	case 'w':			// wait for job to complete
-	    waitForJob = TRUE;
-	    break;
-	case 'x':			// cover page: to's company
-	    proto.setCoverCompany(optarg);
-	    break;
-	case 'y':			// cover page: to's location
-	    proto.setCoverLocation(optarg);
-	    break;
-	case '?':
-	    usage();
-	    /*NOTREACHED*/
-	}
+    switch (c) {
+    case '1':			// restrict to 1D-encoded data
+        proto.setDesiredDF(0);
+        break;
+    case '2':			// restrict to 2D-encoded data
+        proto.setDesiredDF(1);
+        break;
+    case 'a':			// time at which to transmit job
+        proto.setSendTime(optarg);
+        break;
+    case 'b':			// minimum transfer speed
+        proto.setMinSpeed(optarg);
+        break;
+    case 'B':			// desired transmit speed
+        proto.setDesiredSpeed(optarg);
+        break;
+    case 'C':			// cover page: template file
+        proto.setCoverTemplate(optarg);
+        break;
+    case 'c':			// cover page: comment field
+        proto.setCoverComments(optarg);
+        break;
+    case 'D':			// notify when done
+        proto.setNotification("when done");
+        break;
+    case 'd':			// destination name and number
+        addDestination(optarg);
+        break;
+    case 'E':			// disable use of ECM
+        proto.setDesiredEC(FALSE);
+        break;
+    case 'F':			// override tag line format string
+        proto.setTagLineFormat(optarg);
+        break;
+    case 'f':			// sender's identity
+        setFromIdentity(optarg);
+        break;
+    case 'h':			// server's host
+        setHost(optarg);
+        break;
+    case 'I':			// fixed retry time
+        proto.setRetryTime(optarg);
+        break;
+    case 'i':			// user-specified job identifier
+        proto.setJobTag(optarg);
+        break;
+    case 'k':			// time to kill job
+        proto.setKillTime(optarg);
+        break;
+    case 'l':			// low resolution
+        proto.setVResolution(98.);
+        break;
+    case 'M':			// desired min-scanline time
+        proto.setDesiredMST(optarg);
+        break;
+    case 'm':			// medium resolution
+        proto.setVResolution(196.);
+        break;
+    case 'n':			// no cover sheet
+        proto.setAutoCoverPage(FALSE);
+        break;
+    case 'N':			// no notification
+        proto.setNotification("none");
+        break;
+    case 'p':			// submit polling request
+        addPollRequest();
+        break;
+    case 'P':			// set scheduling priority
+        proto.setPriority(optarg);
+        break;
+    case 'r':			// cover sheet: regarding field
+        proto.setCoverRegarding(optarg);
+        break;
+    case 'R':			// notify when requeued or done
+        proto.setNotification("when requeued");
+        break;
+    case 's':			// set page size
+        proto.setPageSize(optarg);
+        break;
+    case 't':			// times to retry sending
+        proto.setMaxRetries(atoi(optarg));
+        break;
+    case 'T':			// times to dial telephone
+        proto.setMaxDials(atoi(optarg));
+        break;
+    case 'v':			// verbose mode
+        verbose++;
+        setvbuf(stdout, NULL, _IOLBF, BUFSIZ);
+        SendFaxClient::setVerbose(TRUE);	// type rules & basic operation
+        FaxClient::setVerbose(verbose > 1);	// protocol tracing
+        break;
+    case 'V':			// cover sheet: voice number field
+        proto.setCoverVoiceNumber(optarg);
+        break;
+    case 'w':			// wait for job to complete
+        waitForJob = TRUE;
+        break;
+    case 'x':			// cover page: to's company
+        proto.setCoverCompany(optarg);
+        break;
+    case 'y':			// cover page: to's location
+        proto.setCoverLocation(optarg);
+        break;
+    case '?':
+        usage();
+        /*NOTREACHED*/
+    }
     if (getNumberOfJobs() == 0) {
-	fprintf(stderr, "%s: No destination specified.\n",
-	    (const char*) appName);
-	usage();
+        fprintf(stderr, "%s: No destination specified.\n",
+            (const char*) appName);
+        usage();
     }
     if (optind < argc) {
-	for (; optind < argc; optind++)
-	    addFile(argv[optind]);
+        for (; optind < argc; optind++) {
+            addFile(argv[optind]);
+        }
     } else if (getNumberOfPollRequests() == 0) {
-	copyToTemporary(fileno(stdin), stdinTemp);
-	addFile(stdinTemp);
+        copyToTemporary(fileno(stdin), stdinTemp);
+        addFile(stdinTemp);
     }
     fxBool status = FALSE;
     fxStr emsg;
     if (callServer(emsg)) {
-	status = login(NULL, emsg)
-	      && prepareForJobSubmissions(emsg)
-	      && submitJobs(emsg);
-	if (status && waitForJob) {
-	    if (getNumberOfJobs() > 1)
-		printWarning("can only wait for one job (right now),"
-		    " waiting for job %s.", (const char*) getCurrentJob());
-	    jobWait(getCurrentJob());
-	}
-	hangupServer();
+        status = login(NULL, emsg)
+            && prepareForJobSubmissions(emsg)
+            && submitJobs(emsg);
+        if (status && waitForJob) {
+            if (getNumberOfJobs() > 1) {
+                printWarning("can only wait for one job (right now),"
+                    " waiting for job %s.", (const char*) getCurrentJob());
+            }
+            jobWait(getCurrentJob());
+        }
+        hangupServer();
     }
-    if (!status)
-	printError(emsg);
+    if (!status) printError(emsg);
     return (status);
 }
 
@@ -230,7 +230,7 @@ void
 sendFaxApp::usage()
 {
     fxFatal("usage: %s [options] [files]\n"
-	"(Read the manual page; it's too complicated)", (const char*) appName);
+        "(Read the manual page; it's too complicated)", (const char*) appName);
 }
 
 /*
@@ -242,22 +242,24 @@ sendFaxApp::addDestination(const char* cp)
     fxStr recipient;
     const char* tp = strchr(cp, '@');
     if (tp) {
-	recipient = fxStr(cp, tp-cp);
-	cp = tp+1;
-    } else
-	recipient = "";
+        recipient = fxStr(cp, tp-cp);
+        cp = tp+1;
+    } else {
+        recipient = "";
+    }
     fxStr dest(cp);
     if (db && dest.length() > 0) {
-	fxStr name;
-	FaxDBRecord* r = db->find(dest, &name);
-	if (r) {
-	    if (recipient == "")
-		recipient = name;
-	    dest = r->find(FaxDB::numberKey);
-	}
+        fxStr name;
+        FaxDBRecord* r = db->find(dest, &name);
+        if (r) {
+            if (recipient == "")
+            recipient = name;
+            dest = r->find(FaxDB::numberKey);
+        }
     }
-    if (dest.length() == 0)
-	fatal("Null destination for \"%s\"", cp);
+    if (dest.length() == 0) {
+        fatal("Null destination for \"%s\"", cp);
+    }
     SendFaxJob& job = addJob();
     job.setDialString(dest);
     job.setCoverName(recipient);
@@ -271,22 +273,23 @@ sendFaxApp::copyToTemporary(int fin, fxStr& tmpl)
 {
     tmpl = _PATH_TMP "/sndfaxXXXXXX";
     int fd = Sys::mkstemp(tmpl);
-    if (fd < 0)
-	fatal("%s: Can not create temporary file", (char*) tmpl);
+    if (fd < 0) {
+        fatal("%s: Can not create temporary file", (char*) tmpl);
+    }
     int cc, total = 0;
     char buf[16*1024];
     while ((cc = Sys::read(fin, buf, sizeof (buf))) > 0) {
-	if (Sys::write(fd, buf, cc) != cc) {
-	    Sys::unlink(tmpl);
-	    fatal("%s: write error", (char*) tmpl);
-	}
-	total += cc;
+        if (Sys::write(fd, buf, cc) != cc) {
+            Sys::unlink(tmpl);
+            fatal("%s: write error", (char*) tmpl);
+        }
+        total += cc;
     }
     Sys::close(fd);
     if (total == 0) {
-	Sys::unlink(tmpl);
-	tmpl = "";
-	fatal("No input data; tranmission aborted");
+        Sys::unlink(tmpl);
+        tmpl = "";
+        fatal("No input data; tranmission aborted");
     }
 }
 
@@ -323,8 +326,7 @@ main(int argc, char** argv)
     signal(SIGTERM, fxSIGHANDLER(sigDone));
     signal(SIGCHLD, fxSIGHANDLER(SIG_DFL));     // by YC
     app = new sendFaxApp;
-    if (!app->run(argc, argv))
-	sigDone(0);
+    if (!app->run(argc, argv)) sigDone(0);
     signal(SIGHUP, fxSIGHANDLER(SIG_IGN));
     signal(SIGINT, fxSIGHANDLER(SIG_IGN));
     signal(SIGTERM, fxSIGHANDLER(SIG_IGN));
